+++ backend/src/main/java/com/optimize/elykia/core/entity/stock/StockTontineRequest.java
@@ -25,6 +25,8 @@ public class StockTontineRequest extends Auditable<String> {
 
     private String collector; // Le commercial
 
+    private String note;
+
     private LocalDate requestDate;
     
     private LocalDate validationDate;